Minor improvements to make cache loaders friendlier
Like CacheLoader, an AsyncCacheLoader can now throw a checked exception
that is automatically rethrown as unchecked by the cache. This avoids
that boilerplate logic in the implementation code.

A small convenience static factory, bulk(mappingFunction), is provided
on the loader interfaces. This delegates singular and bulk loads to
the given function.